Reassign team to User entity
